Exact search for taxonomic names

Search taxonomic names with or without parameters specified.

If parameters are given, the intersection of matching records are returned.

If no parameters are given, all records are returned.

If parameters are specified with NULL specified, examples of valid combinations of values are returned.

am_search_exact(
  SpeciesID = NULL,
  SpecCode = NULL,
  Genus = NULL,
  Species = NULL,
  FBname = NULL,
  OccurRecs = NULL,
  OccurCells = NULL,
  StockDefs = NULL,
  Kingdom = NULL,
  Phylum = NULL,
  Class = NULL,
  Order = NULL,
  Family = NULL,
  deepwater = NULL,
  angling = NULL,
  diving = NULL,
  dangerous = NULL,
  m_invertebrates = NULL,
  highseas = NULL,
  invasive = NULL,
  resilience = NULL,
  iucn_id = NULL,
  iucn_code = NULL,
  iucn_version = NULL,
  provider = NULL
)

Arguments

SpeciesID

AquaMaps unique identifier for a valid species used by the Catalogue of Life Annual Checklist (www.catalogueoflife.org). Example for the whale shark: Fis-30583

SpecCode

Species identifier used in FishBase or SeaLifeBase

Genus

Genus name of the species

Species

Specific epithet of the species

FBname

Common name suggested by FishBase or SeaLifeBase

OccurRecs

Number of point records used to generate good cells

OccurCells

Number of good cells used to generate species envelope

StockDefs

Distribution of the species as recorded in FishBase or SeaLifeBase

Kingdom

Kingdom to which the species belongs

Phylum

Phylum to which the species belongs

Class

Class to which the species belongs

Order

Order to which the species belongs

Family

Family to which the species belongs

deepwater

Does the species occur in the deep-sea (i.e. tagged bathypelagic or bathydemersal in FishBase or SeaLifeBase)? 0=No, 1=Yes

angling

Is the species a sport fish (i.e. tagged as a GameFish in FishBase)? 0=No, 1=Yes

diving

Is the species found on a dive (i.e. where DepthPrefMin in HSPEN < 20 meters)? 0=No, 1=Yes

dangerous

Is the species dangerous (i.e. tagged as traumatogenic or venonous in FishBase or SeaLifeBase)? 0=No, 1=Yes

m_invertebrates

Is the species a marine invertebrate? 0=No, 1=Yes

highseas

Is the species an open ocean fish species (i.e. tagged as pelagic-oceanic in FishBase)? 0=No, 1=Yes

invasive

Is the species recorded to be invasive (i.e. in FishBase or SeaLifeBase)? 0=No, 1=Yes

resilience

Resilience of the species (i.e. as recorded in FishBase/SeaLifeBase)

iucn_id

IUCN species identifier

iucn_code

IUCN Red list classification assigned to the species

iucn_version

IUCN version

provider

FishBase (FB) or SeaLifeBase (SLB)?

Value

tibble with results...

See also

Examples

if (FALSE) { # \dontrun{
am_search_exact()
am_search_exact(Species = "bucculentus")
am_search_exact(FBname = NULL, provider = NULL)
} # }