Total count for number of distinct species across cells, ie "richness"

Global richness map data (i.e. number of different species per cell), optionally filtered for taxonomy and a preferred probability threshold

am_species_per_csc(csc, min_prob = 0.5, keys = NULL)

Arguments

csc

a vector of strings with CsquareCodes (grid cell identifiers)

min_prob

a numeric with the threshold for probable occurrence, default 0.5

keys

a vector of string with SpeciesID values

See also

Examples

if (FALSE) { # \dontrun{
# distinct number of species in a specific grid cell
am_species_per_csc("7516:236:1", 0.99)

# distinct species count per grid cells in LME 1
csc <- am_hcaf() |> filter(LME == 1) |> collect() |> pull(CsquareCode)
am_species_per_csc(csc, 0.9)

# global richness map data based on taxonomy
keys <- am_search_exact(Family = "Carangidae")$SpeciesID
am_species_per_csc(keys = keys, min_prob = 0.9)
} # }